config: T2485: move ifconfig_vlan.py to conf/vlan.py
