+++ src/conf_mode/interfaces-ethernet.py
@@ -21,7 +21,7 @@ from copy import deepcopy
 from netifaces import interfaces
 
 from vyos.ifconfig import EthernetIf
-from vyos.ifconfig_vlan import apply_all_vlans, verify_vlan_config
+from vyos.config.vlan import apply_all_vlans, verify_vlan_config
 from vyos.config.interface import list_diff, intf_to_dict, add_to_dict, interface_default_data
 from vyos.validate import is_member
 from vyos.config import Config
