+++ src/conf_mode/interfaces-pseudo-ethernet.py
@@ -23,7 +23,7 @@ from netifaces import interfaces
 from vyos.config import Config
 from vyos.config.interface import list_diff, intf_to_dict, add_to_dict, interface_default_data
 from vyos.ifconfig import MACVLANIf, Section
-from vyos.ifconfig_vlan import apply_all_vlans, verify_vlan_config
+from vyos.config.vlan import apply_all_vlans, verify_vlan_config
 from vyos import ConfigError
 
 default_config_data = {
